+++ internal/utils/utils.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"log"
 	"os"
+	"strings"
 )
 
 // If the first character is a '!' then assume what follows is a file containing the text
@@ -16,4 +17,30 @@ func OptionallyFromFile(args ...string) string {
 	}
 
 	return string(buffer)
+}
+
+func QuoteAlways(s string) string {
+	s = strings.ReplaceAll(s, "\"", "\"\"")
+
+	return "\"" + s + "\""
+}
+
+func RemoveBlankLines(s string) string {
+	return strings.ReplaceAll(s, "\n\n", "\n")
+}
+
+func QuoteIfNeeded(s string) string {
+	commaIndex := strings.Index(s, ",")
+	tabIndex := strings.Index(s, "\t")
+	quoteIndex := strings.Index(s, "\"")
+
+	if commaIndex == -1 && tabIndex == -1 && quoteIndex == -1 {
+		return s
+	}
+
+	if quoteIndex != -1 {
+		s = strings.ReplaceAll(s, "\"", "\"\"")
+	}
+
+	return "\"" + s + "\""
 }